requirements.txt
numpy
pytorch
tqdm